requirements_integrations.txt
youtube-transcript-api
pytube
notion-client
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gitpython
